paddle-ocr-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
paddleocr==2.7.0
paddlepaddle==2.5.2
numpy==1.25.2
pillow==10.0.0
pydantic==2.4.2 